backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
langchain==0.1.0
langchain-core==0.1.10
langchain-community==0.0.12
langgraph==0.0.20
pydantic
httpx==0.26.0
llama-cloud
llama-index-core
llama-index
python-dotenv==1.0.0
sse-starlette==1.8.2
PyPDF2==3.0.1
chromadb
sentence-transformers
packaging==23.2
chardet
yfinance
